run denylist check pre db insert
