+++ denylist/Cargo.toml
@@ -22,3 +22,4 @@ xorf = {workspace = true}
 serde = {workspace = true}
 serde_json = {workspace = true}
 config = {workspace = true}
+chrono = { workspace = true }
